+++ gui/brassica-gui/mainwindow.h
@@ -87,6 +87,8 @@ private slots:
     void saveLexicon();
     void saveLexiconAs();
 
+    void toggleCursor();
+
     void rulesModified();
     void lexiconModified();